+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -1,3 +1,4 @@
+//@@author tbhbhbh
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
@@ -9,7 +10,8 @@ import java.util.Arrays;
 import org.junit.Test;
 
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.person.PersonContainsBirthdayPredicate;
+import seedu.address.model.person.PersonContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
 
@@ -17,18 +19,22 @@ public class FindCommandParserTest {
 
     @Test
     public void parse_emptyArg_throwsParseException() {
-        assertParseFailure(parser, "     ", String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "     ",
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
     }
 
     @Test
-    public void parse_validArgs_returnsFindCommand() {
+    public void parse_validArgs_returnsSearchCommand() {
         // no leading and trailing whitespaces
-        FindCommand expectedFindCommand =
-                new FindCommand(new NameContainsKeywordsPredicate(Arrays.asList("Alice", "Bob")));
-        assertParseSuccess(parser, "Alice Bob", expectedFindCommand);
+        FindCommand expectedSearchCommand =
+                new FindCommand(new PersonContainsKeywordsPredicate(Arrays.asList("Alice", "friends")));
+        assertParseSuccess(parser, "Alice friends", expectedSearchCommand);
 
         // multiple whitespaces between keywords
-        assertParseSuccess(parser, " \n Alice \n \t Bob  \t", expectedFindCommand);
-    }
+        assertParseSuccess(parser, " \n Alice \n \t friends  \t", expectedSearchCommand);
 
+        // single birthday month represented by 2 digits
+        expectedSearchCommand = new FindCommand(new PersonContainsBirthdayPredicate("05"));
+        assertParseSuccess(parser, "05", expectedSearchCommand);
+    }
 }
