--- src/test/java/seedu/address/logic/parser/SearchCommandParserTest.java
@@ -1,40 +1,0 @@
-//@@author tbhbhbh
-package seedu.address.logic.parser;
-
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
-import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
-
-import java.util.Arrays;
-
-import org.junit.Test;
-
-import seedu.address.logic.commands.SearchCommand;
-import seedu.address.model.person.PersonContainsBirthdayPredicate;
-import seedu.address.model.person.PersonContainsKeywordsPredicate;
-
-public class SearchCommandParserTest {
-
-    private SearchCommandParser parser = new SearchCommandParser();
-
-    @Test
-    public void parse_emptyArg_throwsParseException() {
-        assertParseFailure(parser, "     ",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, SearchCommand.MESSAGE_USAGE));
-    }
-
-    @Test
-    public void parse_validArgs_returnsSearchCommand() {
-        // no leading and trailing whitespaces
-        SearchCommand expectedSearchCommand =
-                new SearchCommand(new PersonContainsKeywordsPredicate(Arrays.asList("Alice", "friends")));
-        assertParseSuccess(parser, "Alice friends", expectedSearchCommand);
-
-        // multiple whitespaces between keywords
-        assertParseSuccess(parser, " \n Alice \n \t friends  \t", expectedSearchCommand);
-
-        // single birthday month represented by 2 digits
-        expectedSearchCommand = new SearchCommand(new PersonContainsBirthdayPredicate("05"));
-        assertParseSuccess(parser, "05", expectedSearchCommand);
-    }
-}
